Safari-609.1.11.1/LayoutTests/editing/deleting/paste-with-transparent-background-color.html

<!DOCTYPE html>
<html style="background: transparent; text-decoration: none">
<body>
<p id="description">This tests cutting and pasting a content with inline "background: transparent;" into a document with a style rule that overrides this style.
WebKit should preserve the inline style declaration. To manually test, cut and paste "hello world WebKit" below. "world" should not be highlighted in blue.</p>
<style>
.test {
    background: blue;
}
</style>
<div id="container" contenteditable>hello <span class=test style="background: transparent;">world</span> WebKit</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description(document.getElementById('description').textContent);

document.getElementById('container').focus();
document.execCommand('SelectAll', false, null);

Markup.dump(document.getElementById('container'), 'Before cut and paste');

document.execCommand('Cut', false, null);
document.execCommand('Paste', false, null);

Markup.dump(document.getElementById('container'), 'After cut and paste');

</script>
</body>
</html>
