third_party/blink/web_tests/editing/deleting/paste-with-transparent-background-color-expected.txt

This tests cutting and pasting a content with inline "background: transparent;" into a document with a style rule that overrides this style.
WebKit should preserve the inline style declaration. To manually test, cut and paste "hello world WebKit" below. "world" should not be highlighted in blue.

Before cut and paste:
| "<#selection-anchor>hello "
| <span>
|   class="test"
|   style="background: transparent;"
|   "world"
| " WebKit<#selection-focus>"

After cut and paste:
| "hello "
| <span>
|   class="test"
|   style="background-color: transparent;"
|   "world"
| " "
| "WebKit<#selection-caret>"
